requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.6
Django==6.0.3
djangorestframework==3.15.2
django-filter==24.3
djangorestframework_simplejwt==5.4.0
idna==3.11
pillow==12.1.1
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-dotenv==1.0.0
requests==2.32.5
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
